requirements.txt
elasticsearch==8.14.0
elastic-transport==8.13.1
ffmpeg==1.4
Flask==3.0.3
Flask-RESTful==0.3.10
flask-swagger-ui==4.11.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
openai-whisper==20231117
python-dotenv
